[SNHU] - CS300 Mod 3, Assignment complete Task 2
The task one requirement was addressed by the initially provided code. In order
to progress the assignment the second task will be started. The LinkedList
constructor needed to be updated to set the initial values for some of the
class members for task 1 of the Module 3 assignment. In order to address the
task requirement, the default constructor will be updated.

- Update the LinkedList default constructor to initialize the `head` and `tail`
pointers to the initial values.

Signed-off-by: Rienour <R.D.Eichenour@gmail.com>